pyproject.toml
[project]
name = "mealie-restore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "requests-toolbelt>=1.0.0",
]
